[ALPS04717140] spm: add ram_consloe dump for mtk idle
[Detail]
1. Add ram_consloe dump for mtk idle
2. fix wrong type

MTK-Commit-Id: ac03e1411b98364fd6403a8ced01b06ea0d9069e

CR-Id: ALPS04717140
Feature: Deep Idle
(cherry picked from commit d4bbed4848212c6fabeb3619b8ee151c5d86c642)

+++ drivers/misc/mediatek/ram_console/mtk_ram_console.c
@@ -1212,7 +1212,7 @@ u32 aee_rr_curr_cidle_data_val(void)
 	return LAST_RR_VAL(cidle_data);
 }
 
-void aee_rr_rec_cidle_time_val(u32 val)
+void aee_rr_rec_cidle_time_val(u64 val)
 {
 	if (!ram_console_init_done || !ram_console_buffer)
 		return;
